Add support for graphical output of flow definition and logging
The new command line option `--format` is available on suitable
commands where the output can be targeted toward a graphical
representation. With value `graphviz` a graphviz compliant rendering
of the flow will be output. With value `image` and image will be
generated by graphviz and opened on the system.

Additionally, some bug fixes related to introspection of a scope to
get its scope were added.