Proposed change for truthiness test
Here's a proposed change, but I'm not 100% sure this is right. It will
allow an empty list to get passed through. I would expect that we'd be
robust enough on the action provider side to handle that empty list,
but I haven't checked/tested.

So, perhaps see if this really does seem to work for what the goal of
this is.