Updates to handle resuming INACTIVE actions/flows
The main new feature is addition of the resume method on both actions
and flows (as usual, flows version built in terms of the action
version). This will check for the special INACTIVE condition based on
a `code` property in the action's details for a `required_scope`
value. If present, it will walk the user through the process of
creating new consents for the `required_scope` value prior to invoking
the resume method on the flow/action.

Other big changes:

* We've imported a new method of managing tokens / directing the user
for consent. This moves away from the FRL package and uses a direct
implementation on the Globus SDK native client. It also, therefore,
prompts on the command line for cut-and-paste to/from a browser rather
than trying to launch a browser directly. Lessons learned from work on
the Timer CLI.

* We've added methods for managing a session including `whoami`,
`logout` and `revoke`. This is particularly helpful in development
when managing tokens and being able to re-set to a known "ground
state" when issuing tokens, etc. See `globus-automate session --help`
for more info.