Updates to match Flows service updates
Some of these include:

- Renaming of role values and migration of filtering to filter_role
instead of filter_roles (plural). Impacts both SDK and CLI.

- Update documentation to talk about different roles, and also
alternative methods of specifying values in expressions and the new
method of catching timeouts.

- Where identities are entered for roles, allow the Globus Auth
username to be used in place of the full urn with uuid. Thus,
john_doe@state.edu could be used to specify a user with a role rather
than using the UUID. Docs updated to guide people as such.

- A few other cleanups here and there.