Cleanup `tox.ini` to simplify
`{envpython}` isn't needed for any of our testenvs, and yet we are
using it. Remove.

`tox -e black` was missing from CI. Fix that.

`basepython` was being specified in the `flake8` build strangely, and
didn't accord with the pre-existing build configuration. Prior to
this set of changes, `flake8` was running under py3.9 only.
Simply remove this usage. Testing on py3.6 is the correct way to
guarantee py3.6 compatibility, and likewise with other old pythons.

Add a space (for readability) to sphinx requirements. Note that
`python -m sphinx` is the same as `sphinx-build`.