Convert or delete YAML flow examples
Where a YAML flow definition appears to provide _any_ value not
duplicated by another example, convert it to JSON. When it is
duplicative, delete it. When it contains elements that cannot be
properly converted (i.e., extended comments), add supporting files
(e.g., a README). If there is no reason for the definition to be here,
remove it.

Signed-off-by: Ada <ada@globus.org>